requirements.txt
mcp[cli]
ollama
pyyaml

